management daily travel stats

+++ management/views.py
@@ -13,7 +13,7 @@ def show_stats(request):# Create your views here.
 	return render_to_response('management/trip_map.html',{ 'page': 'stats', 'request': request, 'routes': routes, })
 
 def deg2rad(deg):
-		return (deg * math.pi / 180.0);
+		return (float(deg * math.pi) / float(180.0));
 
 def rad2deg(rad):
 		return (rad * 180.0 / math.pi);
